+++ test/mocks/mockQueryManager.ts
@@ -2,8 +2,6 @@ import { QueryManager } from '../../src/core/QueryManager';
 
 import mockNetworkInterface, { MockedResponse } from './mockNetworkInterface';
 
-const defaultReduxRootSelector = (state: any) => state.apollo;
-
 // Helper method for the tests that construct a query manager out of a
 // a list of mocked responses for a mocked network interface.
 export default (...mockedResponses: MockedResponse[]) => {